Move RunSpec and PathId.
