+++ src/main/scala/mesosphere/marathon/state/EnvVarValue.scala
@@ -5,16 +5,6 @@ import com.wix.accord._
 import com.wix.accord.dsl._
 import mesosphere.marathon.plugin
 
-sealed trait EnvVarValue extends plugin.EnvVarValue {
-  val valueValidator = new Validator[EnvVarValue] {
-    override def apply(v: EnvVarValue) =
-      v match {
-        case s: EnvVarString => validate(s)(EnvVarString.valueValidator)
-        case r: EnvVarSecretRef => validate(r)(EnvVarSecretRef.valueValidator)
-      }
-  }
-}
-
 sealed trait EnvVarRef extends EnvVarValue {
   val appValidator: Validator[AppDefinition]
 }
